Add functional force live migrate test
Forced live migration now seems to be failing, as after the movement of
the VM, the allocations are only present on the source host, and not on
the destination host.
The related assertions are commented out and replaced with wrong ones to
make the test pass.

Related-Bug: #1712008
Change-Id: I6856f57426db6f2f49daea86679b50d5d019fe19
(cherry picked from commit 52d732d6bdb78ca5c940112e8dff6eef7f6828e8)