Add functional live migrate test
Live migration now seems to be failing, as after the movement of the VM,
the allocations are both present on the source host, and on the
destination host.
The related assertions are commented out and replaced with wrong ones to
make the test pass.

Change-Id: Ib15d34c8da14dd006a0782e325d3c5e7b83fd0f1
Related-Bug: #1712045
(cherry picked from commit 1ce7a136b362fbfb2d3d231ba8e1cd77081139ce)