Add missing unit tests for FilterScheduler._get_all_host_states
This addresses a review comment from change
I8666e0af3f057314f6b06939a108411b8a88d64b for some
missing tests. This code is tightly coupled between
how the SchedulerManager calls FilterScheduler.select_destinations
and how the provider_summaries variable is interpreted differently
when it's None or {}, and is also tightly coupled to how
HostManager._get_computes_for_cells considers compute_uuids being
None or not when grabbing all compute nodes from a cell. Because
of this coupling and complexity, the NOTE in the method is also
updated for clarity.

Change-Id: I3e994d4ad23bf1dd2e46f5d69ff7fdd1a6afdb0e