Add uuid to migration table
This adds a uuid column to the migration table so that in a subsequent
patch we can start using them. Currently we only identify migrations
by internal database id, which is not only a problem for cellsv2, but
also not sufficient for us to use for making allocations that are owned
by a migration process.

Related to blueprint migration-allocations

Change-Id: I21dd6125094b13922b703caa34dffffab1789e1b