How about not logging errors every time we shelve offload?
The delete_allocation_for_instance method does not return
anything and "if not None" is always True, so we're always
logging an error when we shelve offload an instance, which
is not the most fun thing in the world.

This was probably caused by some confusion with how
remove_provider_from_instance_allocation works which does
return a boolean result.

Change-Id: I30b716b0649f6d367ea747149da5c5f921ae19c5
Closes-Bug: #1712689
(cherry picked from commit e208cbb1f122b948af91a8d11c0f45dcec5bbc76)