Delete instance allocations when the instance is deleted
In a deployment that still has Ocata computes, the resource
tracker will delete the allocations for an instance when it
is deleted, but in an all-Pike+ computes deployment, the
allocations will only be removed for a deleted instance
during the update_available_resource periodic task.

The capacity on the node should be freed as soon as the
instance is deleted, like when we shelve offload an
instance, so this does that.

The ServerMovingTests functional tests were glossing over
this by forcing the run of the update_available_resource
periodic task code after the instance was deleted but
before checking the allocations were gone from Placement.
So this change removes that part of the tests to show
this fixes the problem.

Change-Id: I0f57b4ba8e4c3931b70bd737de5b7c8ae47f033d
Closes-Bug: #1712684