libvirt: slow live-migration to ensure network is ready
In Neutron, commit b7c303ee0a16a05c1fdb476dc7f4c7ca623a3f58 introduced
events sent during a live migration when the VIFs are plugged on
destination node.

The Linux bridge agent mechanism driver is detecting new networks on
the destination host only when the TAP devices are created, and these
tap devices are only created when libvirt starts the migration. As a
result, we must actually start the migration and then slow it as we
wait for the neutron events.

This change ensures we wait for these events.

Depends-On: Icb039ae2d465e3822ab07ae4f9bc405c1362afba

Closes-Bug: #1414559
Signed-off-by: Sahid Orentino Ferdjaoui <sahid.ferdjaoui@redhat.com>
Change-Id: I407034374fe17c4795762aa32575ba72d3a46fe8