Change default for [notifications]/default_publisher_id to $host
The default used in code for CONF.notifications.default_publisher_id
is CONF.host, which is a free-form StrOpt.

In an upcoming change, the "my_ip" option is going to be changed from
a StrOpt to an IPOpt, so we made the default value change for
the default_publisher_id option before that.

Change-Id: Ia85130f66bd5a60097e8967064559fd0962d06fd
Closes-Bug: #1713150