Move allocation manipulation out of drop_move_claim()
This only moves the allocation manipulation out of drop_move_claim()
and into a new helper method inside the compute manager. This is
a precursor to changing the behavior in a subsequent patch, so the
intent here is to move it without changing any behavior first.

Related to blueprint migration-allocations

Change-Id: I0883c2ba1989c5d5a46e23bcbcda53598707bcbc