Make allocation cleanup honor new by-migration rules
This makes the cold migration cleanup path play by the "new rules"
about how allocations are handled. In short, if a migration is
confirmed, we delete allocations held under the migration's uuid.
If we're reverting, we prevent the destination compute from manipulating
the allocations at all.

Of course, if the migration uuid was not used to hold an allocation
on the source, then we fall back to the old behavior.

Related to blueprint migration-allocations

Change-Id: I7b2903c56cb53b48afe2c4dec42ae7623a7f24eb