[placement] Add test for empty resources in allocation
If the resources object in an allocation request is an empty dict then
there will be a 500 deep in the belly of the AllocationList object
handling when it tries get a information. The patch demonstrates the bug
with a new gabbi test that verfies the 500. The next patch will fix it.

Change-Id: I5ab417e39044af63fb9acb57c48c087ecf1e6f04
Related-Bug: #1714072
(cherry picked from commit ad2d3fc0f806f46382fa3c0a95c2a5d003d40736)