Refactor out claim_resources_on_destination into a utility
We need this method for a similar fix in evacuate, so this
change pulls the method out of the LiveMigrationTask and moves
it into scheduler utils along with unit tests.

The only functional difference is that instead of raising
MigrationPreCheckError, it now raises NoValidHost, which for
live migration ends up being the same result when it's handled
in ComputeTaskManager.

Change-Id: Ie63a4798d420c39815e294843e02ab6473cfded2
Related-Bug: #1713786
(cherry picked from commit 34f8a351502418d30f137ad49e948802d6b2ab3c)