Add recreate test for evacuate claim failure
This adds a functional recreate test for when the MoveClaim
fails on the destination node and the allocation on the
destination node is not cleaned up. This is because the
MoveClaim fails in it's constructor so it never exits the
Claim context manager to call the drop_move_claim which
would remove the destination node allocation.

Eventually we'll have to manually remove the destination
node allocation in the ComputeManager.rebuild_instance method.

Change-Id: I8900ace4436c4837beb8b4eb1e1d05905efc6dce
Related-Bug: #1713786