Update docs for _destroy_evacuated_instances
The docstring on this method no longer reflects reality,
and hasn't since we started filtering on migration records
starting back in Liberty. This change updates the docstring
for what the method actually does now, and also includes a
comment about why we include 'accepted' but not yet complete
migrations when looking for things to remove from this source
node.

Change-Id: Ia0c67aeffe41e133de9dee5080be9c5ec1739aa6
Related-Bug: #1713783