Remove compatibility code for flavors
The /345_require_online_migration_completion cell database
migration added in Ocata prevents anyone from upgrading
the cell database schema until they have performed the online
flavor data migrations via the nova-manage db online_data_migrations
command.

This means we can remove the compatibility code in the Flavor
object which performs a lookup routine in the API database and
if not found it falls back to the cell database, which at this
point should be empty, therefore making that dead code now.

We can't yet remove the DEPRECATED_FIELDS because that would impact
the object version, so this adds a TODO to remove those when we
create a version 2.0 of this object.

Change-Id: I4107af592448e20fe929f8fa3112929e3685b581