Add nova-manage db command for ironic flavor migrations
This adds `nova-manage db ironic_flavor_migration` which allows
pushing the migrations ahead of time, and without ironic or nova
services running.

This is a similar operation to online_data_migrations, but there is
a need to pass data that would normally be retrieved from ironic.
However, we want this to work offline and thus that data must be
provided locally. When fast-forwarding through the pike release,
ironic and nova-compute will never be started and thus we can't rely
on the online nature of the migration integrated into the ironic
driver.

Conflicts:
      doc/source/cli/nova-manage.rst

NOTE(mriedem): Conflict was due to online_data_migrations not
being documented in Pike.

Change-Id: I459bce472227ee08230d3120d9e2eb1bde635218
(cherry picked from commit 8f8982d8ef381729ef12a14f37486335f31c4fd3)