[placement] Allow _set_allocations to delete allocations
If the used field of an allocation has the value of 0 then use that to
signal that the allocation identified for that consumer and resource
provider is to be deleted within the transaction of _set_allocations.

No explicit deletion is required because _set_allocations deletes all
allocations for a consumer before creating new ones. The trick here is
not creating another one if used is set to 0.

This is going to be used by a future commit that adds support POST
/allocations where it will be possible to add, modify, and delete
allocation records for >1 consumer in one request. At the HTTP API
layer it is currently not possible to provide empty allocations
(to PUT /allocations/{consumer_uuid}) by virtua of JSONSchema. This
will remain true. Only POST will use the functionality added in this
commit.

Change-Id: I585486a240d5c3b8e9ffec7223ca643b331463e6