Fix test_rpc_consumer_isolation for oslo.messaging 5.31.0
With this change in oslo.messaging 5.31.0:

  I0bbf9fca0ecbe71efa87c9613ffd32eb718f2c0e

The endpoint passed in is going to be checked for a 'target'
attribute and check it's type if set, so looking for it has
to be ignored.

There are also some services tests that are mocking out what is
eventually an rpc endpoint and those need to specifically
set the target attribute to None otherwise .get() on the Mock
will return another Mock and cause a failure.

Change-Id: Ic12df8b12b1379f902137d49b605daaf0b89801d
Closes-Bug: #1715470