Make ConductorTaskTestCase run with 2 cells
This lays the groundwork for ConductorTaskTestCase to run
with two cells which is needed for an upcoming fix.

As a result, some mocks needed to be removed for targeting
a cell context, but we still assert a targeted context is
used where necessary, just in a different way. Some other
DB-hitting operations needed to be mocked out in the unshelve
tests since the instance being worked on isn't actually in
a database, and these should have probably been mocked out
anyway.

Change-Id: I6548548825c00a71f3d317ceb37be26da312be31
Related-Bug: #1715493