Regenerate and pass configdrive when rebuild Ironic nodes
Previously, the configdrive could only be set when setting
the node's provisioning state to "active". When rebuilding,
the old configdrive was used and therefore was never updated
with latest content.

Since Ironic API microversion 1.35, it is now allowed to provide
a configdrive when setting the node's provisioning state to "rebuild".

Blueprint: rebuild-ironic-config-drive
Related-bug: #1575935
Depends-On: I9a5529f9fa796c75621e9f4354886bf3032cc248
Change-Id: I1f9056f66519b9ca2f4e23143559735f2bff8943