[placement] Removing versioning from resource_provider objects
Remove import of resource_provider from objects/__init__.py

Also remove VERSION and version annotations from the objects in
nova.objects.resource_provider that were never registered in the first
place.

This finishes the de-registration of all the resource provider objects
and attempts to make it clear that for this file, we don't do that.

Implements: bp placement-deregister-objects
Change-Id: Iea390f70035d7f640abf00462cc4b189548c2700