neutron: handle binding:profile=None during migration
The port binding:profile handling in _setup_migration_port_profile
and _update_port_binding_for_instance is not handling when the
binding:profile is set on the port but the value is None, which
results in a NoneType error. Neutron API does not define a format
for this field, or whether or not it will be specified on the port,
and if so, if it's an empty dict or None, so let's be safe and
make sure we handle None.

Because of how many places we access the binding:profile on
a port, this adds a common helper utility.

Co-Authored-By: Eric M Gonzalez <eric@awnix.com>

Change-Id: I564bac88fad6cc47ccbf7425b1ab39899fdc1c2e
Closes-Bug: #1717365
(cherry picked from commit 8ac7be36bedaf0dd3467efc5b5bffdb365b8231b)
(cherry picked from commit c393e490751416f408432ec20e2b1afc1af88af0)