doc: rename the Indices and Tables section
We haven't had a doc wide index or mod index in this
section on the main page for a long time, so it's a
weird title for just having a link to the search page.

This renames the section title, describes the link and
also links to the openstack-wide docs search page.

Change-Id: Idf63e6035841c8ff0ec1aa9bd2994fedf53555b1