Add base implementation for efficient cross-cell instance listing
This adds a new set of routines that can list instances efficiently
across cells while retaining a stable sort ordering. The theory here
is that we do the query to the cells in parallel, and merge sort the
results.

This is not yet plugged into the actual list instances api call
because it lacks some important features (limit, markers) which
are added in subsequent patches.

Change-Id: I3ffed6eb7008df9d728fe9728b368fec8d3434bc