Account for compute.metrics.update in legacy notification whitelist
The compute.metrics.update legacy unversioned notification was missing
from the whitelist because the only test for it was mocking out the
rpc notifier call, which is what validates the entries in the whitelist.

This adds the entry to the whitelist and fixes the test to actually
use the notifier code.

Change-Id: Ie278ca58c371b8e3bd4861e6c86ebcf30c6778ea
Closes-Bug: #1717943