Use improved instance_list module in compute API
This wires non-cellsv1 deployments up to use the new instance_list
efficienter cross-cell listmaster 9000.

Changes to the unit tests are related to mocks that are no longer valid.
Also, I remove tests that did some cell behavior testing through the
compute api, which is all covered in the functional tests for that
module and no longer provide value.

Note that with this, the behavioral wart that all cell0 instances are
sorted before those in other cells is no longer present (other than
for legacy cellsv1 users which continue to use the old code). However,
instances that are pre-scheduling (i.e. still just BuildRequest stage)
still appear before the rest. This is transient and only affects people
with in-flight builds.

Change-Id: I6f71eda5e25a3ccc102f0a77b6c9058de4be4b38