Copy some tests to a cellsv1 mixin
This patch moves some tests to a cellsv1-specific mixin class ahead of
changing how the plumbing in compute api works. Since we will make a hard
cutover to the new instance query method for all non-CellsV1 code, we
pre-duplicate these tests in a mixin here so we can just fix the base test
class when we make that cut. This makes it easier to just gut cellsv1
later instead of having to remove in-test conditionals for cellsv1.

Change-Id: Idec35d2aab591e988dbd10efb513700ce71521e5