Fix a pagination logic bug in test_bug_1689692
This test attempts to list all instances, then list them again with the
first instance as the marker and ensure that the remaining instances
are returned in the page. Now that we are doing queries to cells in
parallel, consecutive unsorted list queries can return things in
different orders as the cells may reply at different times. The fix
in this patch is to ask for results to be sorted, which is the only way
it makes sense. It worked before purely because we were always scanning
the cells linearly and in the same order.

Related-Bug: #1689692
Change-Id: I3ca2a167c902d565c36a5d5dbba1bf1c214aa20b