Add regression test for persisted RequestSpec.retry from failed resize
Commit 74ab427d4796d8a386f84a15cc49188c2a60f8f1 in Newton added
code to persist changes to the RequestSpec during a resize since
the flavor changes.

That change inadvertantly also persisted any failed hosts during
the resize that are stored in the RequestSpec.retry field during
a reschedule.

The problem is that later those persisted failed hosts are rejected
by the RetryFilter, which can be confusing if an admin is trying
to live migrate or evacate the instance to one of those specific
hosts.

This adds a functional regression test to show the failure, which
will be fixed in a separate change that then modifies the assertions.

Change-Id: Ib8a23db838b0bbf2cfb8123cf6aaa39d00ff0640
Related-Bug: #1718512
(cherry picked from commit 89448bea577b30c40ce39185d14fe14f9c61a0c2)