Restore '[vnc] vnc_*' option support
Changing an option's name when the option's group has been changed in a
previous release results in retained support for only the original
group-name combination and not the intermediate combination.

This behavior can be seen in change Ic05c2c836, which effectively
removed support for the options '[vnc] vncserver_proxyclient_address'
and '[vnc] vncserver_listen' without a deprecation period.

Re-add this support by using the 'deprecated_opts' parameter, which
clearly _is_ still useful [1].

[1] https://review.openstack.org/#/c/364260/

Change-Id: I6405a8024b2bb4df067325b3335008865fd4fdf2
Co-Authored-By: Stephen Finucane <sfinucan@redhat.com>
Closes-Bug: #1718545