Ensure instance can migrate when launched concurrently
When we fixed the problem in If7da79356174be57481ef246618221e3b2ff8200
we forgot to modify a specific check in select_destinations().

Since _schedule() is returning the correct number of needed hosts but
we were still using the wrong number of instances to verify, the
conditional in select_destinations() was always incorrect.

Note that we needed to modify test.py and fake driver because:

 - _do_check_can_live_migrate_destination was using CONF.host
 - check_can_live_migrate_destination in the fake driver was
   incorrectly trying to set a None value to the object while
   libvirt fixed that earlier (block_migration=None when the
   user specifies block_migration='auto' in the API)
 - pre_live_migration in the fake driver was not returning
   migrate_data, which is passed through live_migration and
   _post_live_migration to set the migration object status

Change-Id: Iff839f3478ebe77bf3e2c4becbe9b9b62fff5035
Closes-Bug: #1718455