VMware: implement get_inventory() driver method
Implementing the get_inventory() method allow us to report the
aggregated capacity for cpu, disk and memory. The max_unit property for
cpu and memory is set to the maximum vCPU and memory available from
single ESX host. The max_unit property for disk is set to the maximum
free space available on a single datastore.

Change-Id: I7324bae93e5fbb1301c76466575a906434ca5376
Closes-Bug: #1462957