Fix CellDatabases fixture swallowing exceptions
The _wrap_target_cell() helper in the CellDatabases fixture was
accidentally changing the behavior of the target_cell() functionality
that it wraps by swallowing any exceptions raised and always returning
None. That was not intentional, and introduced by copying the exception
handler below a special case and returning None in a finally: context.

This just removes that exception handler to allow it to propagate to
the caller (as it would in production code), but returns (the
intended special case) if nothing is raised.

This also fixes a bit of the instance_list module that was working around
this quirk so that the code is more natural.

Conflicts:
      nova/compute/instance_list.py

NOTE(mriedem): nova/compute/instance_list.py didn't exist in Pike
and isn't needed here, we just need the fixture fix.

Change-Id: Id611678f53a2ad023dd336e30e62d5c3ae9ffd6e
(cherry picked from commit af05d6208b4f95f1d9b882ed376552b75bb87506)