Refactor resource tracker to account for migration allocations
At the end of update_usage_from_instances(), we call the
_remove_deleted_instances_allocations() to clean up any allocations
that placement says are made against our resource provider. It assumes
that any such allocations that aren't in the known list of instances
must be deleted instances and it removes them. However, once we start
allowing migrations to own allocations against us, we need to make
sure we don't delete those prematurely.

This refactors the resource_tracker to call that method with the list
of migrations active for our node, and it avoids deleting allocations
for providers that match one of those known migrations.

Related to blueprint migration-allocations

Change-Id: Ib240cd863cfdbb538f46e7323b4e282fc3de56cb