Remove allocations when unshelve fails on host
When we unshelve an offloaded instance, the scheduler
creates allocations in placement when picking a host.

If the unshelve fails on the host, due to either the
instance claim failing or the guest spawn failing, we
need to remove the allocations since the instance isn't
actually running on that host.

Change-Id: Id2c7b7b3b4abda8a3b878fdee6806bcfe096e12e
Closes-Bug: #1713796