[placement] Limit number of attempts to delete allocations
In _set_allocations a list of allocations is traversed. For each one, we
delete all the allocations for the consumer uuid in the allocation.
Since some of the allocations in the list can have the same consumer
uuid, make a set of those consumers and only delete once.

This is an optimization that will reduce the number of SQL statements
made by _set_allocations when it is used in forthcoming patches that
enable multiple allocation to be POSTed to /allocations.

Change-Id: Id742cc8b07bcb07fb59986c0643d0d0311be97de