Make etree.tostring() emit unicode everywhere
etree.tostring() returns bytes by default in python3, and a string by
default in python2. For sanity, we make it explicitly emit unicode
everywhere. This requires a python2 fix in guest, as the libvirt
bindings for python2 don't accept a unicode string to migrateToURI3()'s
'param' argument, whereas the python3 bindings do.

Change-Id: I85cd9a903fba310b5ae7bedeed118ca4ea98dff6