Don't use mock.patch.stopall
The use of mock.patch.stopall actually will cause errors if underlying
patchers exist and try to clean up correctly. This is a lazy way to do
cleanup and should be avoided. This refactors the one place where this
was used which causes some other poison mechanism to fail later.

Change-Id: I7641f3c69a6e63ccb78563c63a9c855be4d58ec3